minic: add a register to report TX timestamp readiness
After transmitting a frame in software, the LM32 reads out the
timestamp set by hardware. Naturally, this comes after the frame
has been sent, so the LM32 must wait.

Prior to this patch the only thing it could wait for was DMA idle.
However, DMA of the packet completes before the timestamp is ready.
Thus there was a race condition where the LM32 would see an old TS.