altera: generate phase-aligned 200MHz clock
In the FAIR project, we need to reproduce the BuTiS clock.
It runs at 200MHz and should be aligned to the PPS.
This commit adds a 200MHz clock output from the WR ref PLL.

A small core uses Altera's phase compensation feature to align
the 200MHz rising edge with the PPS, by leveraging a third 25MHz
output signal which can be used to detect the 125-200 relationship.