altera phy: at some point tx_disparity_o timing changed => updated
The WR TX path in the endpoint reset disparity to minus.
To do this, it needs the disparity from the transceiver.
Previously the disparity had to be delayed a few cycles.
At some point this must have changed.

Now the core goes back to sending
  0x17C = K28_5 minus
as commas after packet boundaries.